fix: restrict nonword character removal in some cases, and adjust output formatting
a non-trivial number of entries use "<" or ">" for less than or greater than, and that was
previously getting stripped. apparent negative numbers are being removed entirely now. format
changes in the report are designed based on some issues seen in CV, but in one of the cases there's
actually a Unicode emoji in the input, which appears to be some program's idea of a joke
interpretation of excel's '=#error!' value