feat: support consent inclusion/exclusion lists
software now supports yaml globals::consent_inclusion_file and globals::consent_exclusion_file for
each input dataset. files are tested for existence and loaded. either file or both may be excluded
from the configuration file and the corresponding step will be ignored. either or both files can be
empty. subjects in the exclusion list are first removed if specified; subjects in the inclusion list
are then selected from the remainder. if both files are specified, subjects in the phenotype data
that are not present on either list are collected for reporting in the html report. input file
format for both files is: one subject ID per line, no header. information regarding the inclusion
and exclusion of subjects has been added to the output report, though as usual the report itself is
not yet tested.