fix: improve SAS output
haven's SAS output does not transfer factor level information and certain typing data that is needed
for SAS to interpret the output data correctly. this patch falls back to foreign::write.foreign to
emit sas7bdat and an accompanying source .sas file that can be run in SAS to assign levels and
otherwise improve the end user experience.