feat: add categorical_to_numeric variable type
in the SurveyCTO reports, there are variables that are encoded as categoricals but that contain
purely numeric level labels. these are basically useless as categoricals in most cases, but creating
derived variables for them is a pain. the categorical_to_numeric yaml config type is now recognized.
variables in this type are first treated as categoricals, in that they are converted to R factors;
but then, variables of this type are overridden to the internal numeric type based on their assigned
names, and treated in output reporting as numerics. due to this alternate entry point, some upstream
cleaning operations performed on numerics are skipped in favor of categorical level matching, so if
this behavior is not desired, creating derived variables manually should be preferred