fix: revert automatic bounding of numerics; move global negative number removal
We are seeing NAs encoded in numeric data with negative numbers (ie -99).  We initially thought to
convert these to NA via applying a min bound of 0, but we are also applying a global NA replacement
for all numeric values that are negative.  We've adjusted where that happens so that the stdout
basic logging will emit these changes, and we've removed the per-phenotype automatic bound addition
in the config.