feat: add example file and adjust behavior for corner cases
in writing an example file to go with the rst documentation, a few minor corner cases came up. the
schema validator now accepts null type for inclusion and exclusion lists, to match documentation.
update of entries with a leading apparent decimal point to start with "0." now only affects
variables typed as numerics, and permits numeric content with possible unit suffixes. tests are
updated to reflect this corner case and slightly adjusted behavior.