feat: expose unicode conversion as a plaintext linker file
the unicode replacements in process.unicode.characters started out as a tiny replacement set but
expanded to the point where it was somewhat ridiculous to contain the replacements in code,
certainly in the messy format it was in. this is now exposed in the configuration file
inst/unicode_pattern_replacements.tsv. the unicode characters are configured as \uxxxx or \Uxxxxxxxx
style escape sequences and processed by stringi::stri_unescape_unicode. the documentation for
stringi suggests that some of the \Uxxxxxxxx style characters are unspecifiedly unsupported on
windows. we currently don't have any windows users, but that may change in the future.

fix #15