fix: adjust CTO parser to unit increment after repeat and multiple blocks
previously, the CTO parser was starting the next variable number after a repeat or multiple block
based on the total length of created variables to date. that introduces a major problem when
incremental pulls from upstream CTO data leads to different numbers of repeat responses for certain
variables, as the downstream variable indices shift. that's not intrinsically an issue for the
variables themselves, but causes any derived variables based on those later indices to be
irreparably broken. this patches the behavior such that the downstream index is no longer responsive
to the number of repeats in upstream variables

BREAKING CHANGE: this function will no longer generate identical configuration files for inputs from
previous runs. furthermore, any derived variables based on those prior configs are likely to be
fundamentally incompatible with these new configs.