fix: respect capitalization in CTO's handling of categorical values
though process.phenotypes converts everything to lowercase before analysis, SurveyCTO respects input
case. this led to discrepancies in multiple response variables, when the predicted response header
for a onehot with a capitalized alphabetical categorical *value* (as opposed to label) was being
capitalized in the observed csv header but not in the predicted yaml names. this is now resolved.
alphabetic categorical values are rare and undesirable, but nevertheless apparently accepted by CTO
based on their presence in certain live form documents. initial encoding of harmonized TAG#####
variable names anticipated positive integral categories, and so there are now instances of
HW00001_LVL harmonized variable in certain corner cases. there's nothing intrinsically wrong with
that, and it doesn't seem to break downstream logic for the moment, but it is certainly a deviation
from the format in the rest of the existing config yamls, and may be a target for modification later
on.