style: return tables and plots from markdown helper functions
these helper functions were initially based on code directly from the report rmd, and as such
executed direct cat and print commands on kables and ggplots. for unit testing purposes, these
structures are now returned from the functions and printed/catted separately, such that the type and
structure of the objects can be probed in the test suite.